+++ packages/CHANGELOG.md
@@ -9,6 +9,7 @@
 * IListViewComponentProps now generic to enable types on props.records
 * ListView.fields prop now optional
 * IListViewComponentProps.records -> results (to be consistent with IModelOperationResult)
+* UI Components now support pass-through of standard properties (e.g. style)
 
 ### 0.14.0 - 18th March 2018